chore: move audio type from top level
